pyproject.toml
[project]
name = "ionity-prices"
version = "0.1.0"
description = ""
authors = [
    {name = "Ralf Sigmund",email = "ralf.sigmund@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.39.0,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pymongo[srv] (>=4.15.5,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
